Adjust spacings on the recipes page
This changes the spacings on the recipes page
between the recipe stats and all recipes sections.
Those two sections are also now split to separate
components to make the page component more readable.
