+++ src/styles/globals.css
@@ -13,6 +13,10 @@ body {
   min-height: 100vh;
 }
 
+.Aquarium-PageHeader {
+  padding-bottom: 0;
+}
+
 /* Make the table pagination responsive with the layout. */
 .Aquarium-Pagination {
   display: flex;
